Move api to it's own app
