+++ webapp/accounts/urls.py
@@ -30,7 +30,4 @@ urlpatterns = (
     url(r'^welcome/$', views.WelcomeView.as_view()),
     url(r'^overview/$', views.OverView.as_view(), name='overview'),
     url(r'^profile/$', views.EditProfileView.as_view(), name='profile'),
-
-    url(r'^api/accounts(.json)?$', views.AccountsAPIJSONView.as_view()),
-    url(r'^api/accounts.csv$', views.AccountsAPICSVView.as_view()),
 )
