+++ webapp/accounts/views.py
@@ -14,7 +14,6 @@ from accounts.models import EmailConfirmation, VokoUser, PasswordResetRequest
 from django.conf import settings
 import log
 from ordering.core import get_or_create_order
-from utils import CSVResponse, JSONResponse
 
 
 class LoginView(AnonymousRequiredMixin, FormView):
@@ -189,37 +188,3 @@ class EditProfileView(LoginRequiredMixin, UpdateView):
         messages.add_message(self.request, messages.SUCCESS,
                              "Je profiel is aangepast.")
         return super(EditProfileView, self).form_valid(form)
-
-
-class AccountsAPIView(LoginRequiredMixin, View):
-    def get_raw_data(self):
-        data = []
-        users = VokoUser.objects.all()
-        for user in users:
-            field = {
-                'created_date': user.created.date(),  # rounded to day
-                'is_active': user.is_active,
-                'is_asleep': user.is_asleep,
-            }
-
-            email_confirmation = user.email_confirmation
-            if email_confirmation.is_confirmed:
-                field['confirmed_date'] = email_confirmation.modified.date()
-
-            paid_orders = user.orders.filter(paid=True).order_by("modified")
-            first_paid_order = paid_orders.first()
-            if first_paid_order:
-                field['first_order_date'] = first_paid_order.modified.date()
-
-            data.append(field)
-        return data
-
-
-class AccountsAPIJSONView(AccountsAPIView):
-    def get(self, request, *args, **kwargs):
-        return JSONResponse(self.get_raw_data())
-
-
-class AccountsAPICSVView(AccountsAPIView):
-    def get(self, request):
-        return CSVResponse(self.get_raw_data())
