+++ webapp/ordering/views.py
@@ -18,10 +18,6 @@ from ordering.forms import OrderProductForm
 from ordering.mixins import UserOwnsObjectMixin
 from ordering.models import (Product, OrderProduct, Order, Supplier,
                              OrderRound, ProductCategory)
-from pytz import UTC
-from datetime import datetime
-from utils import CSVResponse, JSONResponse
-from accounts.models import VokoUser
 
 
 class ProductsView(LoginRequiredMixin, ListView):
@@ -341,45 +337,3 @@ class OrderSummary(LoginRequiredMixin, UserOwnsObjectMixin, UpdateView):
 
 class SupplierView(LoginRequiredMixin, DetailView):
     model = Supplier
-
-
-class OrdersAPIView(LoginRequiredMixin, View):
-    def get_raw_data(self):
-        now = datetime.now(UTC)
-        data = []
-        order_rounds = OrderRound.objects.all() \
-            .filter(closed_for_orders__lt=now) \
-            .order_by("open_for_orders")
-        for order_round in order_rounds:
-            products = order_round.products.all()
-            suppliers = set()
-            for product in products:
-                suppliers.add(product.supplier.name)
-            paid_orders = order_round.orders.filter(paid=True)
-            ordering_members = set()
-            for order in paid_orders:
-                ordering_members.add(order.user.id)
-            members = VokoUser.objects \
-                .filter(created__lt=order_round.open_for_orders)
-
-            data.append({
-                'open_for_orders_date': order_round.open_for_orders.date(),
-                'number_of_orders': order_round.number_of_orders(),
-                'number_of_ordering_members': len(ordering_members),
-                'number_of_members': members.count(),
-                'total_revenue': order_round.total_revenue(),
-                'number_of_products': products.count(),
-                'numbers_of_suppliers': len(suppliers),
-                'markup_percentage': order_round.markup_percentage
-            })
-        return data
-
-
-class OrdersAPIJSONView(OrdersAPIView):
-    def get(self, request, *args, **kwargs):
-        return JSONResponse(self.get_raw_data())
-
-
-class OrdersAPICSVView(OrdersAPIView):
-    def get(self, request):
-        return CSVResponse(self.get_raw_data())
